flask-activity/requirements.txt
babel==2.9.0
python-dateutil==2.6.0
flask-moment==0.11.0
flask_sqlalchemy==2.4.4